requirements.txt
alembic==1.7.7
APScheduler==3.9.1
attrs==21.4.0
autopage==0.5.0
backports.zoneinfo==0.2.1
cassandra-csv==0.3
cassandra-driver==3.25.0
certifi==2021.10.8
click==8.0.3
cliff==3.10.1
cmaes==0.8.2
cmd2==2.4.1
colorama==0.4.4
colorhash==1.0.4
colorlog==6.6.0
configparser==5.2.0
cycler==0.11.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Monitoring==1.1.2
fonttools==4.33.3
geomet==0.2.1.post1
greenlet==1.1.2
importlib-metadata==4.8.2
importlib-resources==5.7.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.4.2
kneed==0.7.0
Mako==1.2.0
MarkupSafe==2.0.1
matplotlib==3.5.1
numpy==1.21.6
optuna==2.10.0
packaging==21.3
pandas==1.1.5
pbr==5.8.1
Pillow==9.1.0
prettytable==3.2.0
psutil==5.9.0
pyparsing==3.0.8
pyperclip==1.8.2
pyreadline3==3.4.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
SQLAlchemy==1.4.36
stevedore==3.5.0
threadpoolctl==3.1.0
tqdm==4.64.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
wcwidth==0.2.5
Werkzeug==2.0.2
wincertstore==0.2
xgboost==1.6.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.6.0
pytest==6.2.2
-e .
